refactor(font loading, dev mode): improve font loading behavior and development task runner
