+++ packages/genesys-spark-components/package.json
@@ -34,7 +34,7 @@
     "prettier": "prettier --log-level silent --ignore-unknown --write .",
     "prettier-package-json": "prettier-package-json --write ./package.json",
     "stencil": "stencil build --prod",
-    "stencil.dev": "stencil build --dev --watch --serve --no-open --docs",
+    "stencil.dev": "stencil build --dev --watch --serve --no-open",
     "stylelint": "stylelint --fix \"**/*.{css,html,scss}\"",
     "svgo": "svgo -f ./src/components/stable/gux-icon/icons",
     "test": "TZ=UTC stencil test --spec --e2e",
