+++ web-apps/genesys-spark-examples/src/index.html
@@ -1,6 +1,13 @@
+<!doctype html>
 <html lang="en">
   <head>
     <title>Genesys Webcomponents</title>
+    <!-- 
+    Best practice is to include font styles in apps, but here they are removed
+    to validate that JS font detection and loading works correctly.
+    <link href="/webfonts/urbanist.css" rel="stylesheet">
+    <link href="/webfonts/noto-sans.css" rel="stylesheet">
+    -->
     <meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
     <meta charset="UTF-8" />
   </head>
