+++ web-apps/genesys-spark-examples/src/viewer-template.html
@@ -1,5 +1,8 @@
+<!doctype html>
 <html lang="en">
   <head>
+    <link href="/webfonts/urbanist.css" rel="stylesheet" />
+    <link href="/webfonts/noto-sans.css" rel="stylesheet" />
     <meta http-equiv="Content-Type" content="text/html;charset=utf-8" />
     <meta charset="UTF-8" />
